cmd/dinofs: remove known nodes global for testability
All nodes point to their factory, for child node allocation and for
dependencies (inode number generator, metadata store, blob store).

The factory points to root node (for full path implementation) and
to the mentioned dependencies.

The factory records all known nodes, so it also registers a callback
with the metadata store.

Perhaps a bit stretched but should work and I should be able
to test node rollback for all operations in just few commits now.