message: improve documentation, serialize encoder.Encode and decoder.Decode
The serialization of calls for encoding/decoding is, in practice,
not necessary at the moment, because such calls in other packages
in this module are already serialized by other means. But this is
the right place for such synchronization, clients shouldn't have
to care about the internal state of Encoder's and Decoder's.